Relocate (::>) operator to Frontend.Core.
