+++ src/Frontend.hs
@@ -13,7 +13,6 @@ module Frontend
 where
 
 import Control.Monad.Trans.Except
-import GHC.TypeLits (Symbol)
 import Lucid hiding (href_)
 import Network.HTTP.Types
 import Network.Wai
@@ -127,37 +126,6 @@ aulaActions =
   :<|> Backend.api
   :<|> aulaTesting
 
-data Reply
-
--- FIXME could use closed-type families
-
-type family Singular    a :: Symbol
-type family CaptureData a
-
-type instance Singular Comment            = "comment"
-type instance Singular Idea               = "idea"
-type instance Singular IdeaSpace          = "space"
-type instance Singular IdeaVoteValue      = "vote"
-type instance Singular Reply              = "reply"
-type instance Singular SchoolClass        = "class"
-type instance Singular Topic              = "topic"
-type instance Singular UpDown             = "vote"
-type instance Singular User               = "user"
-type instance Singular IdeaJuryResultType = "jury"
-
-type instance CaptureData Comment            = AUID Comment
-type instance CaptureData Idea               = AUID Idea
-type instance CaptureData IdeaSpace          = IdeaSpace
-type instance CaptureData IdeaVoteValue      = IdeaVoteValue
-type instance CaptureData Reply              = AUID Comment
-type instance CaptureData SchoolClass        = SchoolClass
-type instance CaptureData Topic              = AUID Topic
-type instance CaptureData UpDown             = UpDown
-type instance CaptureData User               = AUID User
-type instance CaptureData IdeaJuryResultType = IdeaJuryResultType
-
-infixr 9 ::>
-type (::>) a b = Singular a :> Capture (Singular a) (CaptureData a) :> b
 
 type AulaMain =
        -- view all spaces
